从double更改为float以优化性能
此次提交主要包括将多个类和结构体中的数据类型从double更改为float，以提高性能和减少内存使用。具体更改包括：

- Matrix3X3Tests、Matrix4X4Tests、TestsHelper、Angle结构体、MathsHelper类中的方法参数和局部变量类型更改。
- Matrix4x4d、Vector2d、Vector3d、Vector4d结构体中的字段、属性、方法参数和返回类型更改。
- Box2d结构体、FrameBuffer类、Sampler2d类、WindowRenderer类、Program类、Rasterizer类中的数据类型和函数调用更新。
- 特别地，更新了Matrix4x4d.CreateLookAt方法的参数类型，以及WindowRenderer_Update和WindowRenderer_Render方法的delta参数类型。
- 在多个片段着色器、Rasterizer类中的变量和计算，以及处理用户输入绘制贝塞尔曲线的代码中，从使用double更改为使用float。

这些更改旨在优化图形处理相关的性能和内存使用，特别是在不需要双精度浮点数提供的额外精度时。

+++ Mathematics/Maths.Tests/Matrix3X3Tests.cs
@@ -10,7 +10,7 @@ public class Matrix3X3Tests
     [DataRow(90)]
     [DataRow(180)]
     [DataRow(360)]
-    public void GenerateRotationMatrix(double degrees)
+    public void GenerateRotationMatrix(float degrees)
     {
         Angle angle = Angle.FromDegrees(degrees);
 
@@ -34,7 +34,7 @@ public class Matrix3X3Tests
     [DataRow(-0.4, -0.3)]
     [DataRow(-0.5, -0.5)]
     [DataRow(-1, -1)]
-    public void GenerateScaleMatrix(double x, double y)
+    public void GenerateScaleMatrix(float x, float y)
     {
         Vector2d point1 = new(1, 1);
         Vector2 point2 = new(1, 1);
@@ -56,7 +56,7 @@ public class Matrix3X3Tests
     [DataRow(-0.4, -0.3)]
     [DataRow(-0.5, -0.5)]
     [DataRow(-1, -1)]
-    public void GenerateTranslationMatrix(double x, double y)
+    public void GenerateTranslationMatrix(float x, float y)
     {
         Vector2d point1 = new(1, 1);
         Vector2 point2 = new(1, 1);
@@ -78,7 +78,7 @@ public class Matrix3X3Tests
     [DataRow(180, -0.4, -0.3)]
     [DataRow(90, -0.5, -0.5)]
     [DataRow(0, -1, -1)]
-    public void GenerateMultiplesMatrix(double degrees, double x, double y)
+    public void GenerateMultiplesMatrix(float degrees, float x, float y)
     {
         Angle angle = Angle.FromDegrees(degrees);
