+++ Programming Assignment/PA2/Program.cs
@@ -27,27 +27,27 @@ internal class Program
         _rasterizer = new Rasterizer(_windowRenderer, SampleCount.SampleCount4)
         {
             Model = Matrix4x4d.Identity,
-            View = Matrix4x4d.CreateLookAt(new(0.0, 0.0, 5.0), new(0.0, 0.0, 0.0), new(0.0, 1.0, 0.0))
+            View = Matrix4x4d.CreateLookAt(new(0.0f, 0.0f, 5.0f), new(0.0f, 0.0f, 0.0f), new(0.0f, 1.0f, 0.0f))
         };
 
-        Vertex a = new(new(2.0, 0.0, -2.0), color: new(1, 0, 0, 1));
-        Vertex b = new(new(0.0, 2.0, -2.0), color: new(0, 1, 0, 1));
-        Vertex c = new(new(-2.0, 0.0, -2.0), color: new(0, 0, 1, 1));
-        Vertex d = new(new(3.5, -1.0, -5.0), color: new(185 / 255.0, 217 / 255.0, 238 / 255.0, 1));
-        Vertex e = new(new(2.5, 1.5, -5.0), color: new(185 / 255.0, 217 / 255.0, 238 / 255.0, 1));
-        Vertex f = new(new(-1.0, 0.5, -1.0), color: new(185 / 255.0, 217 / 255.0, 238 / 255.0, 1));
+        Vertex a = new(new(2.0f, 0.0f, -2.0f), color: new(1, 0, 0, 1));
+        Vertex b = new(new(0.0f, 2.0f, -2.0f), color: new(0, 1, 0, 1));
+        Vertex c = new(new(-2.0f, 0.0f, -2.0f), color: new(0, 0, 1, 1));
+        Vertex d = new(new(3.5f, -1.0f, -5.0f), color: new(185 / 255.0f, 217 / 255.0f, 238 / 255.0f, 1));
+        Vertex e = new(new(2.5f, 1.5f, -5.0f), color: new(185 / 255.0f, 217 / 255.0f, 238 / 255.0f, 1));
+        Vertex f = new(new(-1.0f, 0.5f, -1.0f), color: new(185 / 255.0f, 217 / 255.0f, 238 / 255.0f, 1));
 
         vbo = _rasterizer.CreateVertexBuffer([a, b, c, d, e, f]);
         ibo = _rasterizer.CreateIndexBuffer([0, 1, 2, 3, 4, 5]);
     }
 
-    private static void WindowRenderer_Render(double delta)
+    private static void WindowRenderer_Render(float delta)
     {
         ImGui.Begin("PA 2");
         {
             Vector2 size = ImGui.GetContentRegionAvail();
 
-            _rasterizer.Projection = Matrix4x4d.CreatePerspectiveFieldOfView(Angle.FromDegrees(45), size.X / size.Y, 0.1, 100.0);
+            _rasterizer.Projection = Matrix4x4d.CreatePerspectiveFieldOfView(Angle.FromDegrees(45), size.X / size.Y, 0.1f, 100.0f);
 
             _rasterizer.SetViewport(0, 0, (int)size.X, (int)size.Y);
